+++ src/main/java/net/moonlightflower/wc3libs/misc/Raster.java
@@ -1,10 +1,10 @@
 package net.moonlightflower.wc3libs.misc;
 
-import java.util.Arrays;
-
 import net.moonlightflower.wc3libs.dataTypes.app.Bounds;
 import net.moonlightflower.wc3libs.dataTypes.app.Coords2DF;
 import net.moonlightflower.wc3libs.dataTypes.app.Coords2DI;
+
+import java.util.Arrays;
 
 public abstract class Raster<T> implements Boundable {
 	private Bounds _bounds;
@@ -150,9 +150,9 @@ public abstract class Raster<T> implements Boundable {
 		Size size = getSize();
 		Size otherSize = other.getSize();
 		
-		int minX = (((int) otherCenter.getX()) - otherSize.getWidth() / 2) - (((int) center.getX()) - size.getWidth() / 2);
+		int minX = (otherCenter.getX() - otherSize.getWidth() / 2) - (center.getX() - size.getWidth() / 2);
 		int maxX = minX + otherSize.getWidth() - 1;
-		int minY = (((int) otherCenter.getY()) - otherSize.getHeight() / 2) - (((int) center.getY()) - size.getHeight() / 2);
+		int minY = (otherCenter.getY() - otherSize.getHeight() / 2) - (center.getY() - size.getHeight() / 2);
 		int maxY = minY + otherSize.getHeight() - 1;
 
 		for (int y = minY; y <= maxY; y++) {
